pyproject.toml
[tool.poetry]
name = "reddit"
version = "0.1.0"
description = ""
authors = ["Ihor Horobets <ihor@horobets.me>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
pillow = "^6.2"
psycopg2-binary = "^2.8"

[tool.poetry.dev-dependencies]
django-extensions = "^2.2"
ptpython = "^2.0"
isort = "^4.3"
mypy = "^0.740.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
factory_boy = "^2.12"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
